requirements.txt
SocksiPy_branch==1.01
telethon==1.21.1
